Build source package
