+++ disttools/MANIFEST.in
@@ -1,0 +1,7 @@
+include librflxlang.gpr
+recursive-include src *.ad?
+recursive-include support *.ad? *.gpr
+recursive-include gnatcoll-bindings/iconv *.ad? *.gpr *.c
+recursive-include gnatcoll-bindings/gmp *.ad? *.gpr *.c
+recursive-include python *.py
+recursive-include langkit/support *.gpr *.ad?
