+++ mypy.ini
@@ -12,6 +12,9 @@ ignore_missing_imports = True
 [mypy-setuptools.*]
 ignore_missing_imports = True
 
+[mypy-wheel.*]
+ignore_missing_imports = True
+
 [mypy-langkit.*]
 ignore_missing_imports = True
