+++ requirements.txt
@@ -1,0 +1,11 @@
+black ==20.8b1
+flake8 >=3.9.2, <3.10
+isort >=5.9.1, <5.10
+mypy ==0.910
+pylint >=2.9.3, <2.10
+pytest >=6.2.4, <6.3
+pytest-cov >=2.12.1, <2.13
+pytest-xdist >=2.3.0, <2.4
+wheel
+virtualenv
+./contrib/langkit
